+++ client/src/views/node.js
@@ -23,6 +23,8 @@ exports.nodeInfo = async ({ info, peers, conf: { expert } }) => {
       process.env.BUILD_TARGET != 'web' ? a('.btn.btn-secondary.btn-sm', { attrs: { href: 'settings.html', rel: 'external' }}, 'Server settings') : ''
     , ' '
     , a('.btn.btn-secondary.btn-sm', { attrs: { href: '#/channels' }}, 'Channels')
+    , ' '
+    , a('.btn.btn-secondary.btn-sm', { attrs: { href: '#/deposit' }}, 'Deposit')
     ])
   , expert ? yaml(info) : ''
   ])