+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Add `--pairing-url` CLI option
 
+- bugfix: Properly parse boolean args in RPC console
+
 ## 0.2.3 - 2019-01-31
 
 - docker: Update c-lightning to v0.6.3