+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 
 - ui: Add a new interface for making on-chain deposits
 
+- Update npm dependencies
+
 ## 0.2.3 - 2019-01-31
 
 - docker: Update c-lightning to v0.6.3