+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 0.2.4 - 2019-02-17
 
 - Enable cookie file by default, persists random login credentials to `~/.spark-wallet/cookie`.